// src/types.ts
export interface ListItem {
  key?: string;
  title: string;
  desc: string;
  imageUrl?: string;
}